submissions/2869-telegram-bot/requirements.txt
python-telegram-bot[ext]==20.7
requests>=2.28,<3.0
python-dotenv>=1.0
